refactor: update logger import path to align with recent module restructuring
