+++ src/backend/base/langflow/services/cache/factory.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 
 from langflow.services.cache.service import AsyncInMemoryCache, CacheService, RedisCache, ThreadingInMemoryCache
 from langflow.services.factory import ServiceFactory
-from langflow.utils.logger import logger
+from langflow.logging.logger import logger
 
 if TYPE_CHECKING:
     from langflow.services.settings.service import SettingsService
