+++ src/backend/base/langflow/utils/util.py
@@ -12,7 +12,7 @@ from langflow.schema import Data
 from langflow.services.deps import get_settings_service
 from langflow.template.frontend_node.constants import FORCE_SHOW_FIELDS
 from langflow.utils import constants
-from langflow.utils.logger import logger
+from langflow.logging.logger import logger
 
 
 def unescape_string(s: str):
