+++ src/backend/tests/unit/test_logger.py
@@ -2,7 +2,7 @@ import pytest
 import os
 import json
 from unittest.mock import patch
-from langflow.utils.logger import SizedLogBuffer
+from langflow.logging.logger import SizedLogBuffer
 
 
 @pytest.fixture
